Rename rule and files
