+++ packages/rulesets/src/spectral/functions/reserved-resource-names-model-as-enum.ts
