+++ packages/rulesets/src/spectral/test/reserved-resource-names-model-as-enum.test.ts
